added font and styles for sidebar list
