+++ app/index.html
@@ -2,7 +2,7 @@
 
 <head>
     <title> Beamer </title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" content="initial-scale=1, maximum-scale=1" />
     <meta charset="utf-8">
 </head>
