+++ app/styles/importer.scss
@@ -3,6 +3,7 @@
 $icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
 @import "~bootstrap-sass/assets/stylesheets/bootstrap";
 @import "~angular-material/angular-material";
+@import "../assets/fonts/heiti.ttf";
 @import url("https://fonts.googleapis.com/css?family=Roboto:400,300");
 @import url("https://fonts.googleapis.com/css?family=Lato:400,300");
 @import url("https://fonts.googleapis.com/css?family=Montserrat:400,700");
@@ -13,7 +14,7 @@ $icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
 @import "partials/base";
 @import "partials/landing";
 @import "partials/alerts";
-@import "partials/buttons";
+@import "partials/sidebarButtons";
 @import "partials/forms";
 @import "partials/icons";
 @import "partials/sidebar";
