+++ app/styles/modules/_variables.scss
@@ -5,11 +5,12 @@ $font-lato: 'Lato',
 sans-serif;
 $font-roboto: 'Roboto',
 sans-serif;
-$font-montserrat: "Montserrat",
+$font-montserrat: 'Montserrat',
 "Helvetica Neue",
 Helvetica,
 Arial,
 sans-serif;
+$font-heiti: 'STHeiti';
 // COLORS
 $pink: #D16982;
 $dark-pink: darken($pink, 5%);
