+++ app/styles/partials/_loginButtons.scss
@@ -1,0 +1,74 @@
+//TODO: change to material buttons
+
+/* hollow btn*/
+
+.bm-ghost-btn {
+    @extend %bm-auth-btn;
+    background: transparent;
+    border: remify(2) solid;
+    border-color: #FFF;
+    color: white;
+    &:hover,
+    &:active {
+        border: none!important;
+        background: white!important;
+        color: $purple!important;
+    }
+    &:focus {
+        background: transparent;
+        border: remify(3) solid;
+        border-color: #FFF;
+        color: white;
+    }
+}
+
+.collapsing {
+    -webkit-transition: none;
+    transition: none;
+}
+
+
+/* solid pink btn*/
+
+.bm-solid-btn {
+    @extend %bm-auth-btn;
+    background-color: $pink;
+    color: white;
+    border: none;
+    &:hover {
+        background: $dark-pink;
+        color: white;
+    }
+    &:focus {
+        @extend .bm-solid-btn;
+        background: $dark-pink;
+    }
+}
+
+.bm-solid-btn[disabled] {
+    background-color: $pink;
+    &:hover {
+        background-color: $pink;
+    }
+}
+
+
+/* auth button */
+
+%bm-auth-btn {
+    @extend .text-center;
+    text-align: center;
+    border-radius: 0;
+    font-size: remify(15.6);
+    letter-spacing: remify(3.57);
+    font-family: $font-montserrat;
+    text-transform: uppercase;
+    font-weight: 700;
+    /* remove blue glow */
+    &:focus,
+    &:active:focus,
+    &.active:focus {
+        @extend %bm-no-glow;
+    }
+    @include size(remify(133), remify(55.56));
+}
