+++ app/styles/partials/_sidebarButtons.scss
@@ -1,0 +1,10 @@
+.sidebarBtn {
+    border-radius: 0;
+    background: $purple;
+    color: white;
+    text-transform: none!important;
+    width: 100%;
+    padding: 0!important;
+    margin: 0!important;
+    font-family: $font-heiti;
+}
