+++ app/views/profile/profile-controller.js
@@ -1,5 +1,9 @@
 var ProfileController = function($scope, $http, user, $mdSidenav) {
     console.log('IN PROFILE CONTROLLER', user);
+    this.user = {
+        username: user.username,
+        email: user.email
+    };
     this.$mdSidenav_ = $mdSidenav;
 };
