+++ app/views/profile/profile-view.html
@@ -1,34 +1,25 @@
-<!-- <side-bar>
-    <ul class="bm-sidebar-list">
-        <li>username</li>
-        <li>butter@nutter.com</li>
-        <li>delete account</li>
-    </ul>
-</side-bar>
-<div id="bm-profile-content" class="snap-content" style="background:white">
-    <button ng-click="profileCtrl.toggleSidebar()">Toggle sidebar</button>
-    <div class="bm-profile-icon">
-        <img src="assets/images/profile_icon.svg">
-    </div>
-    <h4>I am The Title Now</h4>
-    <p>
-        <route-list-item routes="profileCtrl.userRoutes"></route-list-item>
-    </p>
-</div> -->
-<meta name="viewport" content="initial-scale=1, maximum-scale=1, user-scalable=no" />
 <div layout="horizontal" ng-cloak layout-fill>
-    <md-sidenav class="site-sidenav md-sidenav-left md-whiteframe-z3" md-component-id="left" md-is-locked-open="$mdMedia('gt-sm')">
-        <md-toolbar class="md-whiteframe-z1">
-            <h1>Users</h1>
-        </md-toolbar>
+    <side-bar>
         <md-list>
             <md-list-item>
                 <md-button>
-                    I am an entry
+                    {{profileCtrl.user.email}}
+                </md-button>
+            </md-list-item>
+            <md-list-item>
+                <md-button>
+                    {{profileCtrl.user.username}}
+                </md-button>
+            </md-list-item>
+            <md-list-item>
+                <md-button>
+                    Delete User
                 </md-button>
             </md-list-item>
         </md-list>
-    </md-sidenav>
+    </side-bar>
+    <!-- 
+     -->
     <div flex layout="vertical" tabIndex="-1" role="main" class="md-whiteframe-z2">
         <md-content flex>
             <md-button hide-gt-sm ng-click="profileCtrl.toggleList()">
